Update install-modes.md
Removed a few commas, and changed setup to 'set up'.

Setup is one word when it is a noun (e.g., “it was a setup!”) or an adjective (e.g., “follow the setup instructions”). It is two words—set up—when it functions as a verb (e.g., “I’m going to set up the computer”).