Clarify, simplify, fix grammatical issues.
Started cleaning up more of this, beginning with the Protocol operator vs SaaS provider section. There are several unclear sections that I either need help with or that need more discussion. I'm going to dig through the official docs, but if someone can clarify now please do. They are:

Rocket Pool’s staked ETH wrapper, rETH, is the purest in DeFi. We foresee a vibrant ecosystem of integrations ranging from lending markets to run validators more efficiently to composability for productivity.  ***ToDo - fix this. Not clear at all.

If you want to run a node in the network, you can stake as a node for as little or long as you wish. The longer you stake though, the bigger the commission you get as it is generated from the staking rewards earned on your node over time. ***ToDo Question - this sounds like compounding. Is it?

The main commission that node operators are awarded in the Rocket Pool protocol is variable and based on supply (node operators available to stake ETH) and demand (ETH needing to be staked). So it can go from very low to high depending on how many deposits are arriving into the network, their size and how many node operators are available to accommodate these new deposits. Currently the minimum commission is 5% of rewards earned and the maximum is 20%. ***ToDo - this needs clarification. 5-20% of whose rewards? The 16 ETH assigned to is from other non-operator users? Are "rewards" and "commission" synonymous here? If not, what defines a commission vs a reward?

Rocket Pool is a decentralised network of node operators. If you are a non-operater staker, your deposit will be allocated to a node operator who will perform Proof of Stake duties. Node operators are required to stake as much ETH as they are assigned — this means they have skin in the game; they are highly incentivised to perform their duties diligently due to this economic bond.  ***ToDo - this last sentence doesn't follow. The bond is from RPL. Their skin is their own 16 ETH, not the additional ETH assigned to them.