Allow native install of Teku Validator to start
The Teku Validator client by default wants to log to `/home/rp/.local/share/teku/logs`, so the `rp` user's home directory must exist or it will fail to start. I offer this merely as a suggestion and to bring awareness to the problem, I'm not tied to any particular solution. A better fix may be to configure Teku Validator to log somewhere else.

https://docs.teku.consensys.net/en/latest/HowTo/Monitor/Logging/