Add note regarding using ledger hardware wallets
The workflow for ledger hardware wallets is fairly confusing and not really intuitive due to Ledgers Lack of native support for MetaMask, and WalletConnect doesnt seem to work at all between the RP site and the Ledger.   This is what worked for me, as well as someone else Patches and I walked thru it today.