Update solo-staker-migration.md
I went through the process on Goerli until I realized I don't have my mnemonic phrase on hand (stored in a safe place, as it should be) to make all of the required changes. As the prerequisits section works like a checklist, I think it makes sense to add it there.